biod: Don't check feature flags on biod startup
On corp-enrolled devices, feature flags are set to default before user
logs in (on corp-enrolled devices, user is not the owner). This means
that biod will trigger beta to production transition before user logs
in to revert the firmware to default version.

When the user logs in, Chrome sends 'RefetchFeatureState' signal. Biod
checks the status of the flag and if it's enabled, biod will trigger
transition from production to beta.

This effectively makes impossible for user to log in, because after
transition from production to beta (that occurs immediately after the
user logs in), the firmware is transitioned back to production firmware
(because the flag is set to default before user logs in).

Fix this problem by removing flag state check on biod startup. Chrome
always sends 'RefetchFeatureState' signal when user logs in, so there is
no need to check it earlier.

BUG=b:347224521
TEST=FEATURES=test emerge-zork biod
TEST=Enable CrOSLateBootAllowBetaFpmcuFirmware flag and reboot.
     Make sure that the firmware is updated to beta and it's not
     reverted to production firmware on login screen.
TEST=Disable CrOSLateBootAllowBetaFpmcuFirmware flag and reboot.
     Make sure that the firmware is updated to production version.

Change-Id: Ie4c3def692201b00c6776807bcd81ca3648f6102
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5837236
Reviewed-by: Tom Hughes <tomhughes@chromium.org>
Commit-Queue: Patryk Duda <patrykd@google.com>
Tested-by: Patryk Duda <patrykd@google.com>