patchpanel: Reimplement DoH rule update in QoSService
QoSService listens to the changes of DoH providers, resolves them to IP
addresses and insert iptables rules for those address to prioritize the
DoH traffic.

Currently this name resolving are very likely to fail since it does not
check if the device is connected to the network. Also the resolved
results might change on different networks, and thus ideally QoSService
should do the resolving on each network change.

This patch changes the implementation to do update on both IPConfig
updates and DoH provider list change. Also do the query directly over
the wifi interface to the DNS servers instead of relying on routing and
resolv.conf. The main reason is since resolv.conf is controlled by
dnsproxy, when QoSService starts query, it's hard to check if dnsproxy
already gets the update and write it to resolv.conf.

BUG=b:361245395
TEST=tast run $DUT network.QosNetworkControl network.DNSProxyUI*
TEST=Manually checked net.log and everything looks fine.

Change-Id: I4349c639f2480e34d6282272456db4254370c7d9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5839234
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Jason Iman <jasongustaman@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>