installer: Balk if FLAGS_* vars are unset in comparisons.
Doing this to avoid shellcheck complaints in a future commit:
shflags always sets these, and shellcheck knows that. But when we move
argument parsing to Rust shellcheck can't tell if these are set or not.
Pre-handle the shellcheck complaints to keep the future commit a little
simpler.

BUG=b:356344778
TEST=UI install on reven

Change-Id: I79cafd3d6fb27079496e15a0ee71f5e001e419b3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5849677
Tested-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>