libmems: device: Do not get event_fd at construction time
When we call ioctl(IIO_GET_EVENT_FD_IOCTL), even if we are not opening
the event file descriptor, it is held until the process ends, so we can
not set the clocktime for the sensor anymore.

The code is needed when one process needs the event file descriptor
while /dev/iio:device is already opened (by iioservice).
Check if it is needed.

BUG=b:364771170
TEST=Check memsetup can set `boottime` via `current_timestamp_clock`
(after stopping iioservice). Check sensor clock is boottime after boot.

Change-Id: I9d04de650c7e6e3ca1b940aed00af8803927eccc
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5847414
Commit-Queue: Gwendal Grignou <gwendal@chromium.org>
Tested-by: Gwendal Grignou <gwendal@chromium.org>
Reviewed-by: Harvey Yang <chenghaoyang@chromium.org>