patchpanel: GuestIPv6: Retry without src when adding /128 routes
When the uplink address is missing on the uplink interface, the request
to add /128 route with src specifier could fail, which could be critical
to connectivity. This patch fallbacks to try to add the /128 route
without src when this happens. If the route can be added successfully
without src, it limits the defect to only direct communication between
the host and the guest, and the most critical data path is not affected.

BUG=none
TEST=unit

Change-Id: I8881d8453b027fd13f88a9c81ebe1e64f032ae49
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5849958
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Tested-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Taoyu Li <taoyl@chromium.org>