shill: Report cellular device without modem as disabled
With a single, always present cellular device, when an expected
or unexpected modem reset happens, there will be a window
where the underlying modem will not be available. During this
time window device should be reported as disabled.
This is also required for current test fixtures which expects
'enabled' to be false during modem resets.

BUG=b:361345218
TEST=Tested in corsolla with modem resets using mmcli

Change-Id: I09ba9fd4d02e7e5afc3e3b97876913ca6073e28f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5854498
Tested-by: Ujjwal Pande <ujjwalpande@google.com>
Commit-Queue: Ujjwal Pande <ujjwalpande@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>