camera: usb: Add quirk for hotpluggable HDMI input devices
When using CfMs users frequently disconnect the HDMI input cable to end
the presentation instead of pressing the "Stop presenting" button in the
UI. For some HDMI input devices this will disconnect the UVC device.

For devices with this behavior, add a quirk to not restart the camera
service when this happens. When the camera service is restarted, it
causes the in-room camera to become muted and requires that the user
press the unmute button in the UI to resume it.

BUG=b:244709400
BUG=b:354766714
TEST=Hotplugged the HDMI input on an Endeavour while presenting

Change-Id: I54aed9f78ca5d7f7d4cebdf13086fa783867b22e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5859114
Reviewed-by: Shik Chen <shik@chromium.org>
Tested-by: Casper <c22813078@gmail.com>
Commit-Queue: Pablo Ceballos <pceballos@chromium.org>