shill: cellular: Force Auto Connect when user triggers a connect attempt
When auto connect is disabled, and the user triggers a manual connect
that requires a precondition to be met before connecting, use the new
SetForceAutoConnect to trigger an AutoConnect after the precondition is
met. The 3 scenarios in which this new behavior applies are:
- Modem is scanning
- Modems is not yet resistered, or the attach APN is being changed.
- A SIM switch needs to occur before connecting.

BUG=b:328239217
TEST=tast run rusty cellular.ShillConnectToUnknownMno.*
Manual tests:
- switch between a eSIM and pSIM multiple times in the login
screen. Verify cellular auto connects after the SIM is switched.
- reboot the device, and click on connect before the modem is
  registered. Verify the device auto connects after registration.

Change-Id: I3dfd9447688fcd2769a18df089c7767e04cd3201
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5861189
Reviewed-by: Michał Mazur <michamazur@google.com>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>