installer: Add utils for wrapping Command
Parts of this are copied from os_install_service. It doesn't feel quite
reusable enough to go in libchromeos-rs.

We use info level to log commands where we ignore the output to make
sure they're visible in logs / feedback reports.
For commands where the output is captured and processed, we log the
command at debug, and let the caller decide what's important to show.

*Issue with ARM*:
It seems there's an issue with running these unit tests on ARM, where
running nonexistent commands returns an exit code, rather than failing
to launch the process. We hypothesize that this is an artifact of the
emulation layer when running these ARM tests on x86. If this actually
fails on ARM we should know pretty quickly when installs start failing.

BUG=b:356344778
TEST=cargo test
TEST=cros_sdk FEATURES=test -- emerge-arm64-generic chromeos-installer
TEST=cros_sdk FEATURES=test -- emerge-reven chromeos-installer

Change-Id: I2ff4ca179d69c244d8fe7cdbd83c40a2ce8c37ad
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5952636
Reviewed-by: Nicholas Bishop <nicholasbishop@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>
Commit-Queue: Ted Brandston <tbrandston@google.com>