lorgnette: add color depth test
This is part of the WWCB SANE backend test suite. The test verifies that
if a scanner supports multiple bit depths for a given color source, then
the `depth` option must be present and formatted correctly. If a scanner
does not support variable bit depth, then this test is skipped.

BUG=b:346843830
TEST=sane_backend_wwcb_tests --scanner=$SCANNER --
--gtest_filter=OptionDescriptorTest.*

Change-Id: Id9aad089172deb9a448ed661e4fa7a205de905d3
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5870841
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>