diagnostics: Reduce the number of EC commands for fan speed queries
In the past, we sent two EC PwmGetFanTargetRpmCommand for each fan to
query all the fan speed: one to get the number of fans and the other for
reading the target fan speed. We can store the responses to reduce the
number of EC required commands.

BUG=b:299559331
TEST=FEATURES=test emerge-volteer diagnostics
TEST=cros-health-tool telem --category=fan

Change-Id: I08bed858ea94c9728dd131bb94a6916fcc4a3260
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5870883
Reviewed-by: Byron Lee <byronlee@chromium.org>
Commit-Queue: Wei-Luan Wang <weiluanwang@google.com>
Tested-by: Wei-Luan Wang <weiluanwang@google.com>