lorgnette: add ADF justification test
This is part of the WWCB SANE backend test suite. This test checks for
the existence of an adf justification option, and makes sure that if
found, the option is formatted as expected by our spec. If the scanner
does not have an ADF, or if the justification option is not found, then
the test is skipped.

BUG=b:346843830
TEST=sane_backend_wwcb_tests --scanner=$SCANNER --
--gtest_filter=OptionDescriptorTest.*

Change-Id: Ia0869d651c1fe294bd224db6205b929dc2f3e071
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5874319
Commit-Queue: Paul Moy <pmoy@chromium.org>
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Paul Moy <pmoy@chromium.org>