vm_tools/maitred: Add SetUpUser RPC
This RPC will be invoked during Baguette installation
to set up the primary user in the VM with a specified username. The RPC creates a new user (if necessary), adds the user to the requested groups, and creates a linger file for systemd.

In Crostini, Tremplin was responsible for a similar user set up,
but for setting up the user within LXD containers.

BUG=b:360170397
TEST=manually rebuilt termina-dlc with the modified maitred \
     and tested within a Baguette installation

Cq-Depend: chromium:5871114

Change-Id: Ia66f987d0c7c8b17d26a2263dd493c15c1f5730a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5854887
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Zihan Chen <zihanchen@google.com>
Auto-Submit: Saswat Padhi <spadhi@google.com>
Commit-Queue: Saswat Padhi <spadhi@google.com>
Tested-by: Saswat Padhi <spadhi@google.com>