shill: wifi: fix down cast caused unit test error in MockLocalDevice
In WiFiProvider::BringDownDevicesByType, a down cast from LocalDevice to
P2PDevice is done to get the shill_id. However, this down cast is not
possible in mock from MockLocalDevice to P2PDevice as MockLocalDevice
inherit from LocalDevice and does not have correlation with P2PDevice.
Change the type to MockP2PDevice which has the ancestor P2PDevice and
LocalDevice so that no ASAN error is reported in unit test.

BUG=b:366091464
TEST=FEATURES=test emerge-amd64-generic chromeos-base/shill

Change-Id: I1d9c19ee938dc97acad63842186405331b7121eb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5878781
Reviewed-by: Kevin Lund <kglund@google.com>
Commit-Queue: Jintao Lin <jintaolin@chromium.org>
Tested-by: Jintao Lin <jintaolin@chromium.org>