shill: cellular: Skip entitlement anomaly detection on non tethering carriers
Do not trigger the anomaly detector when the entitlement check fails
because tethering is disabled on a carrier. There is no need to gather
logs in this case, since this is expected.

BUG=b:368627162
TEST=Test with a verizon SIM. Verify that a crash report is not
generated when the entitlement check fails.

Change-Id: I3e79455f08f562a4f6e856c5451ac25497e0f50e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5883914
Tested-by: Andrew Lassalle <andrewlassalle@chromium.org>
Auto-Submit: Andrew Lassalle <andrewlassalle@chromium.org>
Commit-Queue: Andrew Lassalle <andrewlassalle@chromium.org>
Reviewed-by: Ujjwal Pande <ujjwalpande@google.com>