vm_tools/concierge: Add SetUpVmUser to TerminaVM
In CL:5854887 we added a Maitred RPC to set up a default user in Baguette VMs. In this CL, we add the host-side changes that would invoke the RPC on guest-side Maitred. Specifically, we add a new DBus method in Concierge that will be called by host-side clients (vmc or installation UI etc.).

The `SetUpVmUser` method in Concierge is similar to the `SetUpLxdContainerUser` method in Cicerone, which invokes a guest-side Tremplin RPC to set up a user within LXD containers.

BUG=b:360170397,b:366260782
TEST=dbus-send on DUT

Change-Id: I76558dd8574fc29915934cd6e14516e1ec02d03e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5873906
Commit-Queue: Saswat Padhi <spadhi@google.com>
Tested-by: Saswat Padhi <spadhi@google.com>
Reviewed-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Zihan Chen <zihanchen@google.com>
Auto-Submit: Saswat Padhi <spadhi@google.com>