shill: Use NiceMock by default in unit tests
By default, when a mock object is instantiate from a mock class
directly (without NiceMock or StrictMock), it will be a NaggyMock class,
which will leave a log when there is an uninteresting call (i.e., the
call to a function which is mocked but no EXPECT_CALL set on it).

This patch sets default_mock_behavior to 0 to use NiceMock by default, to
avoid the log spam by uninteresting calls when running the unit tests.

Notes:
- This won't change the verification in the tests, but only affects the
  behavior when there is an uninteresting call.
- This will only affect the behavior of the objects of mock classes by
  default, but won't affect the objects created from NiceMock,
  NaggyMock, or StrictMock.
- Running shill_unittest with `--gmock_default_mock_behavior=1` can
  "restore" the original behavior.

[1] https://google.github.io/googletest/gmock_cook_book.html#NiceStrictNaggy

BUG=None
TEST=FEATURES="test" emerge-$BOARD shill

Change-Id: Idfc2eda3a41df7b6d37741686e505fdc102758af
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5884777
Commit-Queue: Hugo Benichi <hugobenichi@google.com>
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Auto-Submit: Jie Jiang <jiejiang@chromium.org>