dns-proxy: Switch to different users
The patch updates the users to:
- Controller: dns-proxy
- System proxy: dns-proxy-system
- Default proxy: dns-proxy-user
- ARC proxies: dns-proxy-system

ARC proxies use dns-proxy-system to avoid unexpected routing issue with
binding to an interface as "user" when VPN is connected.

This patch also adds a capability for DNS proxy processes to change
users. However, it also adds restriction to not be able to change users
arbitrarily.

DD: http://go/cros-dns-proxy-rootns

BUG=b:363147822
TEST=tast run <DUT> network.DNSProxy*
TEST=tast run <DUT> network.ResolvConf*

Change-Id: I66fead493bef8574747ed9a716288fad33d5152e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5968801
Reviewed-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Jason Iman <jasongustaman@chromium.org>
Tested-by: Jason Iman <jasongustaman@chromium.org>