shill: Make sure unit test can be run directly with gdb
Currently a few tests require the env var `OUT` set to get another file
in this folder, which should always be the case if the unit test is
invoked by `emerge` or `cros_workon_make`, but it's not working directly
when we want invoke the executable directly with `platform2_test.py`.

This patch changes the logic to use the current executable path if `OUT`
is not set. This relies on the assumption (or fact) that the unit test
executable is also installed to the OUT folder directly.

BUG=None
TEST=FEATURES="test" emerge-$BOARD chromeos-base/shill
TEST=Checked that test can run without setting ENV in
  `/mnt/host/source/src/platform2/common-mk/platform2_test.py`

Change-Id: I3cbc869b742ad19bd1e18459deae37e65c0a26d0
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5882287
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Reviewed-by: Andrzej Ostruszka <andrzejo@google.com>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>