cryptohome: Add password delay policy
This refactors the pinweaver auth block a little to take the auth factor
type into account, via the auth factor metadata. It then selects the
policy based on pin vs password metadata.

A side effect of this change is that now if you pass an invalid or empty
metadata to the block, it will fail. This did fail a number of unit
tests which never bothered to populate the metadata in their test calls,
because before this change the metadata was unused. However, this should
not lead to any new failure "in the wild" because on real factors you
would never have empty metadata here.

BUG=b:172221456
TEST=FEATURES=test emerge-$BOARD cryptohome
TEST=tast tests via CQ+1

Change-Id: I4c4c0aca009f5c15f4cbe93c9869c7ba535c235a
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5892312
Reviewed-by: Betul Soysal <betuls@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>