vm_tools: Add unit tests for ReclaimClient and KillsClient
Adds unit tests for ReclaimClient connection and the handler in
KillsClient. The arguments of KillsClient::new() and
ReclaimClient::new() have been changed from &str to &Path. This
is because NamedTempFile is used in testing, and obtaining the
temporary file path as a string could potentially fail and return
an error. Using Path as an argument is therefore a more straightforward
and reliable approach.

BUG=b:359745812
TEST=CQ

Change-Id: I418c478b09b62ab3e92c6e352d2ba45b85ccafa1
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5889670
Reviewed-by: Dmitry Torokhov <dtor@chromium.org>
Reviewed-by: Shin Kawamura <kawasin@google.com>
Reviewed-by: Kameron Lutes <kalutes@chromium.org>
Tested-by: Eri Sawada <erisawada@google.com>
Commit-Queue: Eri Sawada <erisawada@google.com>