shill: network: Do not reset static NetworkConfig in Stop()
The static config is from Service, is only updated when Service is
selected or the StaticIPConfig on the associated Service is updated,
instead of when Network is Start()-ed, so we should not reset it in
Stop(). There is no guarantee that service will be reselected before
Network::Start() is called, so this may cause Network loses the
StaticIPConfig. This is a regression introduced by
https://crrev.com/c/4904009.

BUG=b:358291897
TEST=FEATURES="test" emerge-$BOARD chromeos-base/shill

Change-Id: Ib0f6514fd05872dca42be2d52a46cbb914f862a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5890311
Tested-by: Jie Jiang <jiejiang@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>