installer: Add test of payload image install
Installing from a "payload image" is used by flexor and minios to
install a version of the OS that isn't actively running. Testing that
it works is somewhat tedious, and occasionally I get a step wrong.

Add a `--payload-image` argument to `xtask test-install` to run that
test automatically.

BUG=b:356344778
TEST=cargo +nightly-2024-02-01 \
          xtask test-install \
          --payload-image chromiumos_test_image.bin \
          chromiumos_test_image.bin
TEST=cargo +nightly-2024-02-01 \
          xtask test-install \
          chromiumos_test_image.bin

Change-Id: I6758b84e5f390a38324ccf98f0b55c128497c889
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5889532
Commit-Queue: Ted Brandston <tbrandston@google.com>
Tested-by: Ted Brandston <tbrandston@google.com>
Reviewed-by: Jeffery Miller <jefferymiller@google.com>