cryptohome: Add metadata to AuthBlock::Derive
This parallels an older change from a year ago that similarly added
AuthFactorMetadata to the Create call. This will be needed to allow the
pinweaver block to handle pin vs password factors in Derive, similarly
to how it needs to handle them in Create already.

BUG=b:172221456
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I13455d9f62ac9c0627d6116727185dc29888c076
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5894446
Commit-Queue: John Admanski <jadmanski@chromium.org>
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>