cryptohome: Add PIN check to the migration in Derive
This wraps the PIN-specific delay schedule migration code in a check
that the auth factor is in fact a PIN factor.

Unfortunately, doing the password migration to pinweaver will be
trickier than this one is, because that's a migration between two
different auth blocks that requires a more disruptive operation. But
that will come later.

BUG=b:172221456
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I15c7948b0328349efc1c0df19cca05cf0aee454f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5895035
Commit-Queue: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Reviewed-by: Zi Lin <lziest@google.com>
Tested-by: John Admanski <jadmanski@chromium.org>