power: powerd reports suspend journeys ending in shutdown due to low power
I had to add a TestApi nested class to the Daemon class so I could get
access to the suspender. This makes it so that I can trigger suspend
using the suspender to make sure it's in the correct state, otherwise
various checks fail if I try to trigger a shutdown directly via
power_supply->NotifyObservers() inside the
ShutDownForLowBatteryInDarkResume test.

BUG=b:357970322
TEST=FEATURES=test emerge-brya chromeos-base/power_manager

Change-Id: I23ab68559d1ac413fb23723bf4040cb2f0e997a9
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5966777
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Eizan Miyamoto <eizan@chromium.org>
Reviewed-by: Sam McNally <sammc@chromium.org>
Tested-by: Eizan Miyamoto <eizan@chromium.org>