odml: Remove Stable from coral_service mojoms
We are still in development stage of coral project, and it has been a
few times that we still want to modify some mojom types. Backward
compatibility isn't an issue until we launch the feature, so dropping
stable definition until we're ready to launch it.

BUG=b:358531135
TEST=build ok

Change-Id: Ied240010200a6c6f4412c512134adcca52483e81
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5901681
Commit-Queue: Howard Yang <hcyang@google.com>
Tested-by: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>