cryptohome: Use the pinweaver wrapper with the TPM auth blocks
This has no immediate effect, but turning on the pinweaver for passwords
feature with this would then have TPM factors trigger a recreate,
converting any such blocks.

This change does not add this feature to the scrypt auth block. Both
because the block is used in some ways which complicate adding the
necessary dependencies to it, and because it's not likely to be useful
anyway as any device that's using the scrypt auth block for passwords
isn't going to have pinweaver support.

BUG=b:172221456
TEST=FEATURES=test emerge-$BOARD cryptohome

Change-Id: I33211f134535ee1f2a96bf0e4c8ae30d235534bb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5910210
Tested-by: John Admanski <jadmanski@chromium.org>
Reviewed-by: Hardik Goyal <hardikgoyal@chromium.org>
Commit-Queue: John Admanski <jadmanski@chromium.org>