shill: clear explict_disconnect_ flag when modem is destroyed
In certain cases we may not receive disconnect complete callback
from MM proxy if the modem itself disappears. In such cases we process
modem destroyed callback and stale value of explicit_disconnect_ is not cleared.
If this is not cleared, we can not bring up the interface next time
EstablishLink is called. This is now visible as we moved to a single
cellular device and no longer delete the cellular device. Earlier this
issue was getting masked due to re-creation of new cellular device
so stale value was getting cleared during deletion of old device.

BUG=b:367894837
TEST=Tested on Corsolla

Change-Id: I8fa9dae8348eb91f19f6efa063deb9b7a53f662c
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5910212
Tested-by: Ujjwal Pande <ujjwalpande@google.com>
Reviewed-by: Eric Caruso <ejcaruso@chromium.org>
Commit-Queue: Ujjwal Pande <ujjwalpande@google.com>
Reviewed-by: Aleksander Morgado <aleksandermj@google.com>