shill: network: Introduce NetworkContext
Move logic related to logging from Network into a dedicated class, so
that this class can be passed to other classes easily. Also fix an issue
that the current logging_tag in Network is initialized as ifname, but
will be set to service name after service is selected.

BUG=b:371904984
TEST=FEATURES="test" emerge-$BOARD chromeos-base/shill
TEST=Manually checked that logs are expected.

Change-Id: I5d508fd9f992be1eb907a12bce1572e54ba5123f
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5912953
Tested-by: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Taoyu Li <taoyl@chromium.org>
Commit-Queue: Jie Jiang <jiejiang@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>