arc: Modify GetVerifiedBootParams
With the submission of https://crrev.com/c/5902483,
CTS results showed failures for eve board. During
the override of GetVerifiedBootParams, we returned
KM_ERROR_INVALID_ARGUMENT if any argument could
not be successfully retrieved.

This CL changes that approach to continue till the end
and only include the parameters in the result if they
are available. If the parameters are not available,
only an error would be logged. Also, the unit-test
is removed likewise.

BUG=b:371901422
TEST=emerge-brya chromeos-base/arc-keymint
TEST=cros_workon_make --board=brya chromeos-base/arc-keymint --test
TEST=Tested manually that CTS passes on eve

Change-Id: Iccc9ea609984e4b415a6b96e06b0b1f754144c89
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5914638
Tested-by: Vaibhav Raheja <vraheja@chromium.org>
Commit-Queue: Yao Li <yaohuali@google.com>
Auto-Submit: Vaibhav Raheja <vraheja@chromium.org>
Reviewed-by: Yao Li <yaohuali@google.com>