power: Add DPMST DDC I2C devices to displays lacking I2C paths.
For non-multi-stream-transport (MST) DisplayPort (DP) connectors and
non-DP connectors, a DDC I2C device is present, either directly or
linked to an underlying I2C device, under the sysfs directory for that
connector. This allows use of DDC/CI to, among other things, control the
brightness of the display. For DP-MST connectors, a virtual I2C device
is created, but is parented to the underlying GPU device. As a result,
displays connected via DP-MST do not have an associated I2C device
within powerd and external display brightness control doesn't work.

Work around this shortcoming by associating any unassociated DPMST I2C
devices with displays on the same underlying GPU device lacking an I2C
device.  Note that, with the exception of external display ambient light
sensor integration, all external displays are treated as a group when
adjusting brightness, so it is only necessary that all I2C devices and
all displays are represented, not that they be correctly associated.

BUG=b:352339868
TEST=On a faffy and a taranza, brightness adjustment works on external
     displays connected over DP-MST

Change-Id: I8724d026ceb4d4c8fd3045504675d2cf7d41fefb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5928483
Tested-by: Sam McNally <sammc@chromium.org>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Commit-Queue: Sam McNally <sammc@chromium.org>