vm_tools: cicerone: reorder fuzzer messages
The container_listener_impl_fuzzer.cc file must handle every input
message type in the ContainerListenerFuzzerSingleAction proto message,
but the code handling these messages was in a semi-random order, making
it difficult to tell whether each message type was handled or not.

Sort the messages in the fuzzer switch statement to match the definition
of the message type in the fuzzer.proto file so it can be matched up one
to one. This reveals that the shader cache methods are missing from the
switch statement; add no-op handlers for these messages so the fuzzer
can't hit the NOTREACHED in the default case.

BUG=b:373195749
TEST=presubmit

Change-Id: Ib24960f51d2b98a48442ffdbaa0b3955f31ec0ff
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5930744
Tested-by: Daniel Verkamp <dverkamp@chromium.org>
Reviewed-by: Junichi Uekawa <uekawa@chromium.org>
Commit-Queue: Daniel Verkamp <dverkamp@chromium.org>