odml: [CORAL] Add title observer interface
Add an title observer interface that can be passed to the Group method.
If it is passed, the backend could return a response with empty titles
first, and update the titles through the observer later. The observer is
currently not used in this CL yet.

BUG=b:370851826
TEST=odml unittests

Change-Id: If4842bd29a3079f60a9681bf4b3527f31a7744d4
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5933812
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Takashi Toyoshima <toyoshim@chromium.org>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>