bpf-mons: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=bpf-mons \
        -n 1 \
        --platform2-commit=44413125cd60d1a9bb829e731145a19d265a4a32

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to arc and
-n set to 10.

BUG=b:370478040
TEST=CQ

Change-Id: I361642d6b64114b8d8d9d26ea026ea9b8877c331
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5932422
Reviewed-by: Nicolas Norvez <norvez@chromium.org>
Owners-Override: James Hawkins <jhawkins@chromium.org>
Reviewed-by: James Hawkins <jhawkins@chromium.org>
Tested-by: James Hawkins <jhawkins@chromium.org>
Commit-Queue: James Hawkins <jhawkins@chromium.org>