cecservice: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=cecservice \
        -n 1 \
        --platform2-commit=44413125cd60d1a9bb829e731145a19d265a4a32

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to arc and
-n set to 10.

BUG=b:370478040
TEST=CQ

Change-Id: I2d585d21847c6b18c8f1039e07ece0e39af28447
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5935531
Reviewed-by: James Hawkins <jhawkins@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Owners-Override: James Hawkins <jhawkins@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>