sepolicy: Add policies for os_install_service
Add necessary SELinux policies for the os_install_service
after enabling ARC on the reven board. Those policy changes
are also compatible with non-arc on reven.

This is mostly a reland of chromium:5923599, but with the
device:blk_file ioctl added on cros_chromeos_install
and cros_rsyslogd to be compatible with non-arc on reven
and to make it easier to handle arc disabling situations.

BUG=b:372743967
TEST=ChromeOS Flex test image was successfully installed
TEST=CQ Pass

Change-Id: Id86f4d69a3251e88876c6fa1e9e209a7bdf8875d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5935532
Reviewed-by: Allen Webb <allenwebb@google.com>
Commit-Queue: TingHao Lin <tinghaolin@google.com>
Tested-by: TingHao Lin <tinghaolin@google.com>