chaps: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=chaps \
        -n 1 \
        --platform2-commit=bbb5e435d775855bc366e37a27b0f42914981fe3

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to cfm-dfu-notification and
-n set to 10.

BUG=b:370478040
TEST=CQ

Change-Id: Ic6e7dd264a3262e68379e6073647ce40c17594fb
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5937491
Tested-by: George Burgess <gbiv@chromium.org>
Owners-Override: Hidehiko Abe <hidehiko@chromium.org>
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>