cros-disks: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=cros-disks \
        -n 1 \
        --platform2-commit=bbb5e435d775855bc366e37a27b0f42914981fe3

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to cfm-dfu-notification and
-n set to 10.

BUG=b:370478040
TEST=CQ

Change-Id: I3b23780291f7bb51c4d4acc18e3a75e6eaefbc8e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5937493
Reviewed-by: Hidehiko Abe <hidehiko@chromium.org>
Owners-Override: Hidehiko Abe <hidehiko@chromium.org>
Tested-by: George Burgess <gbiv@chromium.org>
Commit-Queue: George Burgess <gbiv@chromium.org>