shill: Start DHCP-PD when receive no address from SLAAC
When receiving an IPv6 default route update, which suggests an RA with
non-zero router lifetime, SLAACController will start a short timer. If
within the timeout (currently 2s) no IPv6 address is seen on the
interface, SLAACController will determine that address is not available
through SLAAC, and notify Network to start DHCP-PD instead.

This is to adapt to the potential network-side DHCP-PD deployment
without p-flag.

BUG=b:369964250
TEST=network.DHCPPD*

Change-Id: Ia7f71ddf66daa57871292a87b1d0deff0ec7bc27
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5937058
Tested-by: Taoyu Li <taoyl@chromium.org>
Reviewed-by: Hugo Benichi <hugobenichi@google.com>
Commit-Queue: Taoyu Li <taoyl@chromium.org>