federated: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=federated \
        -n 1 \
        --platform2-commit=40f1528ec9807b89fb29bdbb87d495ae3685bedc

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to disk_updater and
-n set to 15.

BUG=b:370478040
TEST=CQ

Change-Id: I0abce497ffa3ce3e69c64dad45092b3d8eb5ee0d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5941720
Owners-Override: James Hawkins <jhawkins@chromium.org>
Reviewed-by: James Hawkins <jhawkins@chromium.org>
Commit-Queue: James Hawkins <jhawkins@chromium.org>
Tested-by: James Hawkins <jhawkins@chromium.org>