hwdrm-videoproc-ta: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=hwdrm-videoproc-ta \
        -n 1 \
        --platform2-commit=97a4f98645710e007b3897c8b091fddc93a647be

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to heartd/ and
-n set to 15.

BUG=b:370478040
TEST=CQ

Change-Id: I19f1e670075170d16cdc12b9fdeede9ce99e99dd
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5940100
Owners-Override: James Hawkins <jhawkins@chromium.org>
Tested-by: James Hawkins <jhawkins@chromium.org>
Commit-Queue: James Hawkins <jhawkins@chromium.org>
Reviewed-by: James Hawkins <jhawkins@chromium.org>