installer: IncludeBraces LSC
This removes the IncludeBraces workaround from .clang-tidy in this
project, and runs `cros format` to bring files in line with
Chromium's new style.

For more information on this LSC, see go/cros-insert-braces-lsc.

To reproduce _just_ this CL, inside of the chroot, run:
    $ ${toolchain_utils}/py/bin/llvm_tools/insert_braces.py \
        --start-project=installer \
        -n 1 \
        --platform2-commit=97a4f98645710e007b3897c8b091fddc93a647be

To reproduce the stack this was a part of, you can run the same
command, but with `--start-project` set to heartd/ and
-n set to 15.

BUG=b:370478040
TEST=CQ

Change-Id: I2bcfe1ba83d1f27fdc77863f749e1ab2e5435624
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5940107
Commit-Queue: James Hawkins <jhawkins@chromium.org>
Reviewed-by: James Hawkins <jhawkins@chromium.org>
Tested-by: James Hawkins <jhawkins@chromium.org>
Owners-Override: James Hawkins <jhawkins@chromium.org>