power: Add workaround for board ovis
This workaround is for the suspend failure cases which could be waived.
The failure rate is minimal, and is not a permanent issue.
In the lid-close scenario, a retry mechanism is available in ChrmoOS.
Given that this is not a permanent failure and does not impact platform
power significantly and end user experience.

With this workaround, we could test other issues behind the failure.
Also reduce the fail rate in factory build or other autotest if we
apply it in the related toolkit.

BUG=b:360227822
TEST=intel_prepost_suspend --boardwa=ovis
     'iotools mmio_read32 0xfe001b0c'
      0x402

Change-Id: I07c4060cb8fcfb7f7581c3fbad5737cbfe48435e
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5937026
Commit-Queue: Curtis Chen <curtis.chen@intel.com>
Reviewed-by: Puthikorn Voravootivat <puthik@chromium.org>
Tested-by: Curtis Chen <curtis.chen@intel.com>