lorgnette: sane_backend_tests: SANETest.MultipleCancel wait for 1st page
The SANETest.MultipleCancel test does not wait for a page to be ready
before invoking its first start_scan() call. The instructions are also
not explicit about requiring a page to be ready for scanning.

Add a wait for the first page to be scanning and more explicit
instructions.

BUG=b:346843300
TEST=(dut) $ sane_backend_wwcb_tests --scanner='drm260:usb:002:002' --
--gtest_filter='SANETest.MultipleCancel'

Change-Id: I5dec24760314dead3a19591092f8641ae8aec34d
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5939785
Reviewed-by: Bryan Cain <bryancain@chromium.org>
Commit-Queue: Aaron Massey <aaronmassey@google.com>
Tested-by: Aaron Massey <aaronmassey@google.com>