lorgnette: add ScanAreaPageDims test
This test makes sure that the br-x, br-y, tl-x and tl-y options
are present and conform to the WWCB spec. In addition, if the
page-height and page-width options are found, this test makes sure
that the units and types match those of br-x, etc.

BUG=b:346843830
TEST=sane_backend_wwcb_tests --scanner=$SCANNER --
--gtest_filter=OptionDescriptorTest.*

Change-Id: I0f5c7a24ff13b9cefd01f14818d3fbd5b290d452
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5943184
Reviewed-by: Aaron Massey <aaronmassey@google.com>
Tested-by: Paul Moy <pmoy@chromium.org>
Commit-Queue: Paul Moy <pmoy@chromium.org>