odml: [CORAL] Process only 1 request at a time
For the embedding engine and title generation engine, the current
behavior is that all request can be pipelined, only queued and handled
together in the model loading part to avoid duplicate loads. However,
the most time-consuming part (and therefore relevant for the pipeline
behavior) in both engines are model loading and model inference.

For model loading, in concurrent requests case, even if we queue the
request at Engine::Process level, there should be no latency penalty
because the model only has to be loaded once and the second request can
skip the model loading stage entirely.
That leaves nothing to benefit from the concurrent request support
because the model backend only supports 1 inference at a time, so the
end time of the second request should be roughly the same no matter we
deal with the requests concurrently or sequentially.

On the other hand, there is 1 benefit of processing requests 1 by 1:
embedding engine caches the result embedding for same prompt, and we
plan to add title caching for similar groups in the title generation
engine soon. Dealing with requests 1 by 1 instead of concurrently will
increase the cache hit rate because it can utilize the cache of the
first request. This is especially useful given that concurrent requests
will likely have similar/same content (caused by user spamming the
feature trigger). Therefore, change the behavior to queueing requests at
Engine::Process level.

BUG=b:361689713
TEST=odml unittests

Change-Id: I7eafb9862adeffdc8fc1c20f637b0aaf5a634016
Reviewed-on: https://chromium-review.googlesource.com/c/chromiumos/platform2/+/5948263
Commit-Queue: Howard Yang <hcyang@google.com>
Reviewed-by: Yi Chou <yich@google.com>
Tested-by: Howard Yang <hcyang@google.com>